requirements.txt
fastapi==0.95.1 # FAST API framework
uvicorn==0.22.0 # ASGI server
pydantic==1.10.7 # Data validation and settings management using python type annotations
python-multipart==0.0.6 # multipart form support
boto3==1.26.129 # AWS SDK for Python
opensearch-py==2.2.0 # OpenSearch client
python-dotenv==1.0.0 # .env file support
diagrams==0.23.3 # Diagram as code
